% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_datasets.R
\docType{data}
\name{acp_voitures}
\alias{acp_voitures}
\title{acp_voitures}
\format{
A data.frame with car-related variables (rows = observations, columns = variables)
}
\source{
datasets/acp_voitures.csv
}
\usage{
acp_voitures
}
\description{
Example dataset: cars measurements used for demonstration of PCA/clustering
}
\examples{
data(acp_voitures)
head(acp_voitures)
}
\keyword{datasets}
