% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_datasets.R
\docType{data}
\name{AUTOS2005subset}
\alias{AUTOS2005subset}
\title{AUTOS2005subset}
\format{
A data.frame with automobile measurements and categorical features
}
\source{
datasets/AUTOS2005subset.txt
}
\usage{
AUTOS2005subset
}
\description{
Subset of AUTOS2005 dataset (example automobile data)
}
\examples{
data(AUTOS2005subset)
head(AUTOS2005subset)
}
\keyword{datasets}
