% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_helpers.R
\name{create_model}
\alias{create_model}
\title{Create clustering model instance}
\usage{
create_model(algorithm = "kmeans", params = list())
}
\arguments{
\item{algorithm}{Character Algorithm name ("kmeans", "hac", "acm")}

\item{params}{List of algorithm-specific parameters}
}
\value{
An R6 object instance of the specified clustering algorithm
}
\description{
Factory function to create instances of clustering algorithms
}
