% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_datasets.R
\docType{data}
\name{jobrate}
\alias{jobrate}
\title{jobrate}
\format{
A data.frame with job rate related variables
}
\source{
datasets/jobrate.csv
}
\usage{
jobrate
}
\description{
Employment/job rate dataset for examples
}
\examples{
data(jobrate)
head(jobrate)
}
\keyword{datasets}
