% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_datasets.R
\docType{data}
\name{protein_dataset}
\alias{protein_dataset}
\title{protein_dataset}
\format{
A data.frame with protein measurement variables
}
\source{
datasets/protein_dataset.csv
}
\usage{
protein_dataset
}
\description{
Protein measurements example dataset
}
\examples{
data(protein_dataset)
head(protein_dataset)
}
\keyword{datasets}
