% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_datasets.R
\docType{data}
\name{tea_data}
\alias{tea_data}
\title{tea_data}
\format{
A data.frame with 300 observations and 18 categorical variables describing tea consumption patterns:
\describe{
\item{breakfast, tea.time, evening, lunch, dinner, always}{Timing of tea consumption (binary: yes/not)}
\item{home, work, tearoom, friends, resto, pub}{Places where tea is consumed (binary: yes/not)}
\item{Tea}{Type of tea preferred (black, Earl Grey, green)}
\item{How}{How tea is taken (alone, milk, lemon, other)}
\item{sugar}{Sugar usage (sugar, No.sugar)}
\item{how}{Packaging format (tea bag, unpackaged, tea bag+unpackaged)}
\item{where}{Purchase location (chain store, tea shop, chain store+tea shop)}
\item{price}{Price category (p_branded, p_variable, p_upscale, p_private label, p_unknown)}
}
All variables are factors, making this dataset ideal for ACM (Multiple Correspondence Analysis) clustering.
}
\source{
datasets/tea_data.csv
}
\usage{
tea_data
}
\description{
Tea consumption and preferences dataset with categorical variables
}
\examples{
data(tea_data)
head(tea_data)
# Excellent dataset for ACM clustering of categorical variables
\dontrun{
# Example ACM clustering
model_acm <- ClustVarACM$new(K = 3)
model_acm$fit(tea_data)
model_acm$plot(type = "biplot")
}
}
\keyword{datasets}
