% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_datasets.R
\docType{data}
\name{vehicle}
\alias{vehicle}
\title{vehicle}
\format{
A data.frame with vehicle characteristics
}
\source{
datasets/vehicle.csv
}
\usage{
vehicle
}
\description{
Vehicle dataset for clustering demonstrations
}
\examples{
data(vehicle)
head(vehicle)
}
\keyword{datasets}
