% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_datasets.R
\docType{data}
\name{vote_catvarclus}
\alias{vote_catvarclus}
\title{vote_catvarclus}
\format{
A data.frame with categorical variables suitable for ACM
}
\source{
datasets/vote_catvarclus.csv
}
\usage{
vote_catvarclus
}
\description{
Synthetic voting / categorical variables dataset for ACM examples
}
\examples{
data(vote_catvarclus)
head(vote_catvarclus)
}
\keyword{datasets}
